+++ axual-webclient/instance.go
@@ -5,10 +5,18 @@ import (
 	"net/url"
 )
 
-func (c *Client) GetInstancesByAttributes(attributes url.Values) (*InstancesResponseByAttributes, error) {
-	o := InstancesResponseByAttributes{}
+func (c *Client) GetInstanceByName(name string) (*InstanceResponse, error) {
+	o := InstanceResponse{}
+	err := c.RequestAndMap("GET", fmt.Sprintf("%s/instances/search/findByName?name=%s", c.ApiURL, url.QueryEscape(name)), nil, nil, &o)
+	if err != nil {
+		return nil, err
+	}
+	return &o, nil
+}
 
-	err := c.RequestAndMap("GET", fmt.Sprintf("%s/instances/search/findByAttributes?%s", c.ApiURL, attributes.Encode()), nil, nil, &o)
+func (c *Client) GetInstanceByShortName(shortName string) (*InstanceResponse, error) {
+	o := InstanceResponse{}
+	err := c.RequestAndMap("GET", fmt.Sprintf("%s/instances/search/findByShortName?shortName=%s", c.ApiURL, url.QueryEscape(shortName)), nil, nil, &o)
 	if err != nil {
 		return nil, err
 	}