+++ internal/tests/ApplicationPrincipalResource/axual_application_principal_connector_key_trailing_ws.tf
@@ -1,0 +1,6 @@
+resource "axual_application_principal" "connector_axual_application_principal" {
+  environment = axual_environment.tf-test-env.id
+  application = axual_application.tf-test-app.id
+  principal   = "${file("{{CERTS}}/generic_application_1.cer")}\n\n"
+  private_key = "${file("{{CERTS}}/generic_application_1.key")}\n\n"
+}